# src/api/swagger.yaml
basePath: /api/v1
definitions:
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.JobTemplate:
    properties:
      createdAt:
        type: string
      defaults:
        description: Default meta/env as JSON
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      engine:
        description: '"nextflow", "wdl", "script"'
        type: string
      hcl_template:
        type: string
      id:
        type: integer
      is_built_in:
        type: boolean
      name:
        type: string
      schema:
        description: JSON Schema for params validation
        type: string
      updatedAt:
        type: string
      version:
        type: string
    type: object
...
  types.JobAddDto:
    properties:
      pipeline_name:
        type: string
      pipeline_params:
        items:
          type: integer
        type: array
      pipeline_version:
        type: string
      user_id:
        type: integer
    required:
    - pipeline_name
    - pipeline_params
    - pipeline_version
    type: object
  types.JobDeleteDto:
    properties:
      job_id:
        type: integer
      user_id:
        type: integer
    required:
    - job_id
    type: object
  types.JobStopDto:
    properties:
      job_id:
        type: integer
      user_id:
        type: integer
    required:
    - job_id
    type: object
...
host: localhost:8086
info:
  contact:
    name: Antelope Team
  description: Antelope bioinformatics pipeline management platform API.
  license:
    name: MIT
  title: Antelope API
  version: "1.0"
paths:
...
  /job/add:
    post:
      consumes:
      - application/json
      description: Submit a new Nextflow job for the current user
      parameters:
      - description: Job submission payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.JobAddDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Submit job
      tags:
      - jobs
  /job/delete:
    delete:
      consumes:
      - application/json
      description: Delete a job owned by the current user
      parameters:
      - description: Job delete payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.JobDeleteDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete job
      tags:
      - jobs
  /job/details:
    get:
      description: Get details of a job owned by the current user
      parameters:
      - description: Job ID
        in: query
        name: job_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get job details
      tags:
      - jobs
  /job/stop:
    post:
      consumes:
      - application/json
      description: Stop a running job owned by the current user
      parameters:
      - description: Job stop payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.JobStopDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Stop job
      tags:
      - jobs
...
  /templates:
    get:
      description: List job templates
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: List job templates
      tags:
      - templates
    post:
      consumes:
      - application/json
      description: Create job template
      parameters:
      - description: job template payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.JobTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create job template
      tags:
      - templates
  /templates/{id}:
    delete:
      description: Delete job template
      parameters:
      - description: template id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete job template
      tags:
      - templates
    get:
      description: Get job template
      parameters:
      - description: template id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get job template
      tags:
      - templates
    put:
      consumes:
      - application/json
      description: Update job template
      parameters:
      - description: template id
        in: path
        name: id
        required: true
        type: string
      - description: job template updates
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update job template
      tags:
      - templates
...
  /user/job/log:
    get:
      description: Stream a job's log over Server-Sent Events
      parameters:
      - description: Nomad allocation ID
        in: query
        name: allocId
        required: true
        type: string
      - description: Log type (stdout or stderr)
        in: query
        name: logType
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Stream job log via SSE
      tags:
      - jobs
  /user/jobs:
    get:
      description: List jobs for the current user with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List current user's jobs
      tags:
      - jobs
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {accessToken}"
    in: header
    name: Authorization
    type: apiKey
  RefreshAuth:
    description: Enter "Bearer {refreshToken}" — used only for /auth/refresh
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"